feat: Added more fields to models; fix: Changed user edit func

+++ models/additional.go
@@ -10,7 +10,7 @@ type AdditionalInfo struct {
 	gorm.Model
 	CompanyID                uint   `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
 	QuarterID                uint   `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
-	Version                  int    `gorm:"not null;index:idx_unique_comp_quarter_version,unique;default:1"`
+	Version                  uint32 `gorm:"not null;index:idx_unique_comp_quarter_version,unique;default:1"`
 	GrowthChallenges         string // bit 0
 	SupportNeeded            string // bit 1
 	PolicyChanges            string // bit 2
@@ -20,8 +20,8 @@ type AdditionalInfo struct {
 	InitiativeProgress       string // bit 6
 	BusinessModelAdjustments string // bit 7
 
-	IsVisible  int `gorm:"default:255"`
-	IsEditable int `gorm:"default:255"`
+	IsVisible  uint8 `gorm:"default:255"`
+	IsEditable uint8 `gorm:"default:255"`
 }
 
 func (a *AdditionalInfo) TableName() string {
