+++ models/assessment.go
@@ -40,7 +40,7 @@ type SelfAssessment struct {
 	gorm.Model
 	CompanyID         uint       `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
 	QuarterID         uint       `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
-	Version           int        `gorm:"not null;index:idx_unique_comp_quarter_version,unique;default:1"`
+	Version           uint32     `gorm:"not null;index:idx_unique_comp_quarter_version,unique;default:1"`
 	FinancialRating   int        // 1-10 bit 0
 	MarketRating      int        // 1-10
 	ProductRating     int        // 1-10
@@ -50,8 +50,8 @@ type SelfAssessment struct {
 	Priorities        Priorities `gorm:"type:text"`
 	IncubatorSupport  string     // bit 7
 
-	IsVisible  int `gorm:"default:255"`
-	IsEditable int `gorm:"default:255"`
+	IsVisible  uint8 `gorm:"default:255"`
+	IsEditable uint8 `gorm:"default:255"`
 }
 
 func (s *SelfAssessment) TableName() string {
