+++ models/fund.go
@@ -8,9 +8,9 @@ import (
 
 type FundraisingStatus struct {
 	gorm.Model
-	CompanyID             uint `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
-	QuarterID             uint `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
-	Version               int  `gorm:"not null;index:idx_unique_comp_quarter_version,unique;default:1"`
+	CompanyID             uint   `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
+	QuarterID             uint   `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
+	Version               uint32 `gorm:"not null;index:idx_unique_comp_quarter_version,unique;default:1"`
 	LastRound             string
 	CurrentInvestors      string
 	InvestorRelations     string
@@ -19,8 +19,8 @@ type FundraisingStatus struct {
 	InvestorPipeline      string
 	ValuationExpectations string
 
-	IsVisible  int `gorm:"default:127"`
-	IsEditable int `gorm:"default:127"`
+	IsVisible  uint8 `gorm:"default:127"`
+	IsEditable uint8 `gorm:"default:127"`
 }
 
 func (f *FundraisingStatus) TableName() string {
