+++ models/risk.go
@@ -8,9 +8,9 @@ import (
 
 type RiskManagement struct {
 	gorm.Model
-	CompanyID          uint `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
-	QuarterID          uint `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
-	Version            int  `gorm:"not null;index:idx_unique_comp_quarter_version,unique;default:1"`
+	CompanyID          uint   `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
+	QuarterID          uint   `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
+	Version            uint32 `gorm:"not null;index:idx_unique_comp_quarter_version,unique;default:1"`
 	RegulatoryChanges  string
 	ComplianceStatus   string
 	RegulatoryConcerns string
@@ -20,8 +20,8 @@ type RiskManagement struct {
 	KeyDependencies    string
 	ContingencyPlans   string
 
-	IsVisible  int `gorm:"default:255"`
-	IsEditable int `gorm:"default:255"`
+	IsVisible  uint8 `gorm:"default:255"`
+	IsEditable uint8 `gorm:"default:255"`
 }
 
 func (r *RiskManagement) TableName() string {
