+++ models/team.go
@@ -8,9 +8,9 @@ import (
 
 type TeamPerformance struct {
 	gorm.Model
-	CompanyID              uint `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
-	QuarterID              uint `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
-	Version                int  `gorm:"not null;index:idx_unique_comp_quarter_version,unique;default:1"`
+	CompanyID              uint   `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
+	QuarterID              uint   `gorm:"not null;index:idx_unique_comp_quarter_version,unique"`
+	Version                uint32 `gorm:"not null;index:idx_unique_comp_quarter_version,unique;default:1"`
 	TeamSize               string
 	NewHires               string
 	Turnover               string
@@ -20,8 +20,8 @@ type TeamPerformance struct {
 	SkillGaps              string
 	DevelopmentInitiatives string
 
-	IsVisible  int `gorm:"default:255"`
-	IsEditable int `gorm:"default:255"`
+	IsVisible  uint8 `gorm:"default:255"`
+	IsEditable uint8 `gorm:"default:255"`
 }
 
 func (t *TeamPerformance) TableName() string {
